+++ cmd/main.go
@@ -87,7 +87,7 @@ type CLI struct {
 	Browser    string `kong:"short='b',help='Path to browser to open URLs with',env='AWS_SSO_BROWSER'"`
 	ConfigFile string `kong:"name='config',default='${CONFIG_FILE}',help='Config file',env='AWS_SSO_CONFIG'"`
 	Lines      bool   `kong:"help='Print line number in logs'"`
-	LogLevel   string `kong:"short='L',name='level',,help='Logging level [error|warn|info|debug|trace] (default: info)'"`
+	LogLevel   string `kong:"short='L',name='level',help='Logging level [error|warn|info|debug|trace] (default: info)'"`
 	UrlAction  string `kong:"short='u',help='How to handle URLs [open|print|clip] (default: open)'"`
 	SSO        string `kong:"short='S',help='AWS SSO Instance',default='Default',env='AWS_SSO',predictor='sso'"`
 	STSRefresh bool   `kong:"help='Force refresh of STS Token Credentials'"`
@@ -316,7 +316,7 @@ func logLevelValidate(level string) error {
 	case "error", "warn", "info", "debug", "trace", "":
 		return nil
 	}
-	return fmt.Errorf("Invalid value for --log-level: %s", level)
+	return fmt.Errorf("Invalid value for --level: %s", level)
 }
 
 func urlActionValidate(action string) error {